config script bugfix
This fix came as a result of just totally replacing gnome-terminal. It
really isn't the best thing out there. I chose Kitty mostly due to ease
of installation and the fact that it had an already made NORD color
scheme. Actually, the color scheme for Kitty is perfectly accurate to
nord, where as the one I had for gnome-terminal made a couple of
concessions. Anyway, I should eventually figure out how to use git
submodules for the color scheme config since I haven't modified it all
from the repo I *cough* stole it from, and don't intend to make any such
changes. I've also changed the compton config to work with kitty, and
the i3bar colors to match *real* nord colors. I added the config files
to the alt + z guy, and I've changed the edit_config function to be a
little cleaner as well, now expanding ~ in the list